+++ testing/test2.py
@@ -1,2 +1,0 @@
-def get_sun(a,b):
-    return a + b